Some updates

+++ tools/bench/YuvBenchmark10.cpp
@@ -52,7 +52,7 @@ void SparkyuvYCbCr444P10ToRGBA10(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -118,7 +118,7 @@ void LibYuvYCbCr444P10ToRGBA8(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -182,7 +182,7 @@ void SparkyuvYCbCr422P10ToRGBA10(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -248,7 +248,7 @@ void LibYuvYCbCr422P10ToRGBA8(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -312,7 +312,7 @@ void SparkyuvYCbCr420P10ToRGBA10(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -378,7 +378,7 @@ void LibYuvYCbCr420P10ToRGBA8(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -442,7 +442,7 @@ void SparkyuvRGBAP10ToYCbCr420P10(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -494,7 +494,7 @@ void SparkyuvRGBA10ToYCbCr422P10(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
@@ -546,7 +546,7 @@ void SparkyuvRGBA10ToYCbCr444P10(benchmark::State &state) {
   rgbaStride = sizeof(uint8_t) * inWidth * 4;
   rgbaData.resize(rgbaStride * inHeight);
   sparkyuv::RGBToRGBA(inSrcData.data(), inWidth * sizeof(uint8_t) * 3, rgbaData.data(), rgbaStride, inWidth, inHeight);
-  int rgba16Stride = sizeof(uint8_t) * inWidth * 4;
+  int rgba16Stride = sizeof(uint16_t) * inWidth * 4;
   std::vector<uint8_t> rgba16Data(rgba16Stride * inHeight);
   sparkyuv::WideRGBA8To10(rgbaData.data(),
                           rgbaStride,
